Address /audit T3 findings — version bump, tests, cleanup
Single-PR pass on the T3 audit report. All 12 findings handled (8 in
code, 4 deferred with documentation).

## High

### #1 Version bump v1.3.0 / build 7
- `project.yml` `MARKETING_VERSION: 1.2.0 → 1.3.0`,
  `CURRENT_PROJECT_VERSION: 6 → 7`. Submission-blocker — the v1.3 work
  on main was still labelled v1.2.0.

### #2 Test coverage for new services
- 6 new test suites, 20 tests:
  - `URLImporterTests` (5) — invalid URL, scheme rejection (file/http),
    `isVideoURL` extension allowlist.
  - `DeepLinkHandlerTests` (3) — non-wallnetic scheme ignored,
    playPause routing, unknown host fallback.
  - `ErrorReporterTests` (2) — `report` publishes, `surface: false`
    silent log-only path.
  - `SlideshowGeneratorTests` (4) — empty assets, even H.264 dims,
    bitrate scaling, settings defaults.
  - `WallpaperMetadataStoreTests` (3) — round-trip for titles,
    colors, tags through the JSON helpers fixed in #167.
  - `AIGenerateViewModelTests` (3) — initial state, cancel clears
    state, `clearGeneratedVideo`.
- All 57 tests (37 prior + 20 new) pass.

### #3 Sandbox runtime test
- Cannot exercise without Developer ID-signed archive. Documented in
  the security PR (#179) checklist; carry-forward, not a code change.

## Medium

### #4 WallpaperManager facade — protocols defined
- New `Services/WallpaperManagerProtocols.swift` declaring
  `WallpaperReading` (read-only) and `WallpaperWriting` (mutating)
  protocols. `WallpaperManager` conforms via extension.
- New code accepting these protocols can be unit-tested without
  touching the singleton. Migration of the 71 existing `.shared`
  call sites is incremental and tracked separately — protocols are
  in place from day one.

### #5 Engine → Services dependency inversion
- Deferred to v1.4 — touches 4 controllers and 24 call sites,
  bigger than this audit-fix scope. Documented as future work.

### #6 Silent catch triaj
- Real number was 1, not 46 (the original grep counted all
  `} catch {` blocks; most have substantive Log/state-mutation
  bodies). The single genuinely silent catch was in
  `Wallpaper.swift:59` (AVAsset metadata probe failure). Now logs
  at `.debug` so unexpected failures surface in Console without
  spamming production.

### #7 DiscoverView file split
- Extracted `WebViewWrapper` + `Coordinator` (~270 lines) to
  `Views/Main/BrowserWebView.swift`. `DiscoverView.swift` shrinks
  748 → 477 lines.

### #8 CHANGELOG.md
- New `CHANGELOG.md` at repo root in Keep-a-Changelog format
  covering v1.3.0 / v1.2.0 / v1.1.0. README's inline changelog
  block replaced with a one-paragraph summary + link.

## Low

### #9 CollectionsView TODO
- Opened follow-up issue #181. Comment now references the issue
  number instead of `// TODO`.

### #10 Force unwraps removed
- 4 `urls(for:in:).first!` sites converted to a shared
  `applicationSupportURL()` helper in `Utils/Log.swift` that
  `fatalError`s with a clear message instead of crashing on a
  nil unwrap. macOS guarantees the directory exists; the helper
  just makes the invariant explicit.

### #11 xcodeproj malformed group warning
- Investigated. The `SharedWidgetModels.swift` cross-target
  reference triggers the warning by design — fixing it via
  XcodeGen `group:` directive caused a build break for the widget
  target. Left as-is. Xcode build succeeds cleanly.

### #12 `try?` documentation
- The remaining 44 `try?` sites are all legitimate cleanup
  (`removeItem`, `defer handle.close()`) or intentional optional
  reads (`attributesOfItem` for size, AVAsset duration probes).
  Documented in the audit report as P3 follow-up; no per-site
  comments added.

## Build + tests
- Debug ✓ Release ✓
- 57 tests pass

Closes T3 audit findings (#1, #2, #4, #6, #7, #8, #9, #10).
Defers #3 (signing required), #5 (v1.4), #11 (cosmetic), #12 (P3).
